added __all__ to api, and chagned xml decimal -> python Decimal in objectify

+++ pyspreedly/objectify.py
@@ -2,6 +2,7 @@ import xml.etree.ElementTree
 from xml.etree import ElementTree as ET
 from cStringIO import StringIO
 from datetime import datetime
+from decimal import Decimal
 import re
 
 _sub_dash = re.compile('-')
@@ -11,7 +12,7 @@ _types = {
     'string'   :  lambda x: x,
     'integer'  :  int,
     'datetime' :  lambda s: datetime.strptime(s, '%Y-%m-%dT%H:%M:%SZ') if s else None,
-    'decimal'  :  float,
+    'decimal'  :  Decimal,
     'boolean'  :  lambda x: x == 'true',
     'array'    :  lambda x: [],  ## Return an empty array
     }
